% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Track.R
\name{getAverage}
\alias{getAverage}
\title{Get the adjustable running window average of the data}
\usage{
getAverage(i, frames, widths, data)
}
\arguments{
\item{i}{The index within 'frames' at which to calculate an average over a window centered at this location}

\item{frames}{The frames in this track}

\item{widths}{A vector of window widths appropriate for each frame in the 'frames' of this track}

\item{data}{The vector of data for which we will calculate the windowed averages}
}
\description{
Get the adjustable running window average of the data
}

