% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/TrackFitting.R
\name{getBulkPhaseShift2}
\alias{getBulkPhaseShift2}
\title{Fist SSE BULK}
\usage{
getBulkPhaseShift2(trackList, resolution = 0.05)
}
\value{
The best fit of a sweep fitted to all the data simultaneously to obtain an ensemble prediction of the phase shift to be used for all the Tracks
}
\description{
Fist SSE BULK
}

