% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Track.R
\name{getDerivative}
\alias{getDerivative}
\title{Get the derivative of a vector}
\usage{
getDerivative(x, t)
}
\arguments{
\item{x}{A numeric vector on which to calculate the derivative}

\item{t}{A numeric vecotor of times with which to determine dt for derivative calculations}
}
\description{
Get the derivative of a vector
}

