% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/PackageFunctions.R
\name{getSweep}
\alias{getSweep}
\title{Return information and data for a log frequency sweep with the given parameters}
\usage{
getSweep(amplitude = 1, phaseShift = 0, offset = 0, sin = FALSE,
  ti = 0, fi = 2, ff = 0.1, sweepDuration = 300, t = seq(0, 300,
  0.05), guess = NULL)
}
\arguments{
\item{amplitude}{numeric}

\item{phaseShift}{numeric - default=0}

\item{offset}{numeric - defulat=0}

\item{sin}{boolean T=sinsoid and F=triangular - default=FALSE}

\item{fi}{numeric initial frequency of the sweep}

\item{ff}{numeric final frequency of the sweep}

\item{sweepDurtion}{numeric}

\item{tAll}{numeric vector All possible times for this}
}
\description{
Return information and data for a log frequency sweep with the given parameters
}

