% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Track.R
\name{interpolateDerivative}
\alias{interpolateDerivative}
\title{This is a three point interpolation of the derivative where the interpolated point
is the middle of the 3 points. This simplifies to the the three-point midpoint formula
when the time steps are equal but can handle when timesteps are unequal (i.e., the
time-step on either side of the 3 points is not equal)}
\usage{
interpolateDerivative(f0, f1, f2, x0, x1, x2, xj)
}
\arguments{
\item{f0}{numeric left function value}

\item{f1}{numeric middle function value}

\item{f2}{numeric right function value}

\item{x0}{numeric left independent value}

\item{x1}{numeric middle independent value}

\item{x2}{numeric right independent value}
}
\description{
This is a three point interpolation of the derivative where the interpolated point
is the middle of the 3 points. This simplifies to the the three-point midpoint formula
when the time steps are equal but can handle when timesteps are unequal (i.e., the
time-step on either side of the 3 points is not equal)
}

