% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/LAPJV.R
\name{lapjv}
\alias{lapjv}
\title{Code adapted from matlab code developed by Yi Cao at Cranfield University (see code header for more info)}
\usage{
lapjv(costMat, resolution)
}
\arguments{
\item{costMat}{A numeric matrix representing the pairwise cost/distance between all points}

\item{resolution}{A numeric vector The number decimal places to keep during calculations (needed as distances are internally represented as integers for speed)}
}
\value{
list(rowsol=rowsol,cost=cost,v=v,u=u,costMat=costMat) where rowsol = the optimal col indicies associated with each row, cost = the final optimal distance cost, v = column reduction numbers, u = row reduction numbers, costMat = the user provided costMat
}
\description{
Code adapted from matlab code developed by Yi Cao at Cranfield University (see code header for more info)
}

