% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Track.R
\name{localDerivative}
\alias{localDerivative}
\title{Get the local derivative around a point in a vector accounding for boundary scenarios at the start and end of the vector}
\usage{
localDerivative(x, t, i)
}
\arguments{
\item{x}{A numeric vector of data}

\item{t}{A numeric vector of time for calculating dt of the derivative}

\item{i}{A numeric value indicating the index in the x and t for which to calculate the local derivative}
}
\description{
Get the local derivative around a point in a vector accounding for boundary scenarios at the start and end of the vector
}

