% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Maxima.R
\docType{class}
\name{Maxima-class}
\alias{Maxima}
\alias{Maxima-class}
\title{Class for storing and working with a set of points}
\description{
Class for storing and working with a set of points
}
\section{Fields}{

\describe{
\item{\code{frame}}{A length one numeric vector}

\item{\code{points}}{A data.frame storing the id, x, and y information of each point in the set of points}
}}
\section{Methods}{

\describe{
\item{\code{addPoint(id, x, y)}}{#' Add a point to this Maxima object}

\item{\code{getPoint(index = NULL, id = NULL)}}{#' Get a point at a specific index or with a specific id value}

\item{\code{getXYZ()}}{#' Return the X, Y, and Z locations for all the points in this Maxima object. If Z is not given, returns 0 for Z.}

\item{\code{initializeWithROI(frame, polygon = NULL)}}{#' initialize this with information from a semicolon separated list of comma separated point information (i.e., 'x,y,id;x,y,id;x,y,id;...')}

\item{\code{length()}}{#' Return the number of points in this Maxima object}

\item{\code{plotMaxima(IDs = TRUE, pch = 20, xlab = "X [pixel]", ylab = "Y [pixel]",
  cex = 0.5, ...)}}{#' Plot the x, y locations of the points in this Maxima object with point ids as labels}
}}

