% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/MaximaList.R
\docType{class}
\name{MaximaList-class}
\alias{MaximaList}
\alias{MaximaList-class}
\title{A class for storing and working with list of Maxima objects. This class is the
class that actually does the point tracking over time.}
\description{
A class for storing and working with list of Maxima objects. This class is the
class that actually does the point tracking over time.
}
\section{Fields}{

\describe{
\item{\code{maxima}}{A list of Maxima objects}

\item{\code{maxID}}{The maximum ID encountered in all the Maxima in the MaximaList}

\item{\code{trackBackStart}}{A single numeric value indicating the frame at which cells were tracked backward from (i.e., usually a late frame)}

\item{\code{trackBackEnd}}{A single numeric value indicating the frame at which cells were tracked back to (i.e., usually an early frame)}
}}
\section{Methods}{

\describe{
\item{\code{generateMaximaPlots(path = NULL, baseName = "Frame_")}}{#' Produce plots of the maxima locations over time and output them to a folder}

\item{\code{getIndexOfFrame(frame)}}{#' Get the index of a frame with a particular number}

\item{\code{getMaxima(frame)}}{#' Get the maxima at a particular frame number (vs index)}

\item{\code{getMaximaAtIndex(index)}}{#' Get the maxima at a particular index (vs frame number)}

\item{\code{getPreviousMaxima(frame)}}{#' Get the maxima that preceeds the given frame (good for backtracking)}

\item{\code{getProp(fun = function(.maxima) {     return(base::range(.maxima$points$x)) })}}{#' Get the property of a maxima by applying the user-specified function on each maxima object}

\item{\code{getTrackList(sin = FALSE, fi, ff, sweepDuration, t0_Frame = 0,
  timePerFrame = 0.035, frameRange = c(0, -1))}}{#' Create a TrackList object from this MaximaList object (essentially a list indexed/grouped by cell id instead of time)}

\item{\code{initializeWithJEXROIFile(path, timeDimName = "Time")}}{#' Initialize with a file (eg., .jxd ROI file from JEX) that stores all the point sets as semicolon separated comma separated point information (see Maxima documentation).}

\item{\code{length()}}{#' Return the number of sets of maxima in this MaximaList object}

\item{\code{offsetFrames(offset = 0)}}{#' Add the offset to all the frame numbers, adjusting trackBaskStart and trackBackEnd accordingly}

\item{\code{setMaxima(newMaxima)}}{#' Put a maxima into the MaximaList object}

\item{\code{trackBack(startFrame, endFrame, maxDist = 150, direction = c(1, 0, 0),
  directionality = 10, uniformityDistThresh = 3, digits = 1)}}{#' Track backwards from startFrame (usually from a later point in time of the acquistion) to endFrame (earlier in the acquisition)}
}}

