% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/PackageFunctions.R
\name{reorganizeTable}
\alias{reorganizeTable}
\title{Take an arff file and reorganize it into a more standard 'table' format. Specifically this is used to
import an arff file from JEX as JEX uses a column called 'Measurement' to define the type of measurment
or property being stored and 'Value', the value of that property.}
\usage{
reorganizeTable(data, baseName = NA, convertToNumeric = TRUE,
  nameCol = "Measurement", valueCol = "Value")
}
\arguments{
\item{data}{An object that is the result of using foreign::read.arff(file) on an arff file}

\item{baseName}{An optional basename to add to whatever label is in the \code{nameCol} portion of each row entry}

\item{convertToNumeric}{An option to convert the columns of information within \code{data} leading up to
\code{nameCol} and \code{valueCol} to numeric or to leave as text. Default is to convert to numeric (i.e., TRUE)}

\item{nameCol}{The name of the column that describes the nature of the value in the \code{valueCol}}

\item{valueCol}{The name of the column with the values of the properties listed in the \code{nameCol}}
}
\description{
Take an arff file and reorganize it into a more standard 'table' format. Specifically this is used to
import an arff file from JEX as JEX uses a column called 'Measurement' to define the type of measurment
or property being stored and 'Value', the value of that property.
}

