% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/TrackFilters.R
\name{trackFrameFilter}
\alias{trackFrameFilter}
\title{Filter for testing if tracks exist within indicated frame bounds.}
\usage{
trackFrameFilter(track, startMin = 0, startMax = 1e+06, endMin = 0,
  endMax = 1e+06)
}
\arguments{
\item{track}{Track The track to be tested}

\item{startMin}{numeric The min (inclusive) frame for the starting frame of the track - default=0}

\item{startMax}{numeric The max (inclusive) frame for the starting frame of the track - default=1000000}

\item{endMin}{numeric The min (inclusive) frame for the ending frame of the track - default=0}

\item{endMax}{numeric The max (inclusive) frame for the ending frame of the track - default=1000000}
}
\value{
boolean whether the track meets the specified conditions
}
\description{
Filter for testing if tracks exist within indicated frame bounds.
}

