% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/TrackFilters.R
\name{trackLengthFilter}
\alias{trackLengthFilter}
\title{Filter for testing tracks based on a min and max length}
\usage{
trackLengthFilter(track, min = 0, max = 1e+06)
}
\arguments{
\item{track}{Track The track to filter}

\item{min}{numeric The minimum length of a track (inclusive, >=)}

\item{max}{numeric The maximum length of a track (inclusive, <=)}
}
\value{
boolean whether the track meets the specified conditions
}
\description{
Filter for testing tracks based on a min and max length
}

